exfat-utils: make -n option fallthrough to -l option for backward compatibility with old utils
Vojtech Trefny concern about backwards compatibility with the fuse exfat
userspace utilities.
The difference is in command line options for mkfs.exfat that are using -l
for label, the exfat fuse mkfs is using -n.
This patch make -n option fallthrough to -l option for backward
compatibility with old utils.

Reported-by: Vojtech Trefny <vtrefny@redhat.com>
Signed-off-by: Namjae Jeon <linkinjeon@gmail.com>